step2: 세부 구현 사항 및 테스트 수정
